+++ wiringPi/wiringPiI2C.h
@@ -29,12 +29,12 @@ extern "C" {
 extern int wiringPiI2CRead           (int fd) ;
 extern int wiringPiI2CReadReg8       (int fd, int reg) ;
 extern int wiringPiI2CReadReg16      (int fd, int reg) ;
-extern int wiringPiI2CReadRegN(int fd, int reg, int *value, uint32_t N); // Erez Added 
+extern int wiringPiI2CReadRegN(int fd, int reg, int *value, uint8_t N); // Erez Added 
 
 extern int wiringPiI2CWrite          (int fd, int data) ;
 extern int wiringPiI2CWriteReg8      (int fd, int reg, int data) ;
 extern int wiringPiI2CWriteReg16     (int fd, int reg, int data) ;
-extern int wiringPiI2CWriteRegN(int fd, int reg, int *value, uint32_t N); // Erez added 
+extern int wiringPiI2CWriteRegN(int fd, int reg, int *value, uint8_t N); // Erez added 
 
 extern int wiringPiI2CSetupInterface (const char *device, int devId) ;
 extern int wiringPiI2CSetup          (const int devId) ;
